ice: Sync driver to dev-queue
Several updates to the ice driver have already been submitted to next-queue
(via intel-wired-lan mailing list). These will eventually be submitted to
net-next.  However, as this RFC is being submitted to the RDMA tree, it's
necessary to "catch up" the ice driver in the RDMA tree to the latest code
in next-queue to ensure functional correctness.

So this patch is a consolidation of all ice driver changes that are in
next-queue but not in the RDMA tree and is being submitted here only for
the purposes of this RFC.

Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>