ice: Register RDMA peer devices to the virtual platform bus
Instead of using basic bus primitives which was created and
managed by LAN driver, attach the peer RDMA devices with
the OS provided virtual platform bus. The peer RDMA driver
independently registers as platform driver and will bind to
the devices.

This model is inspired by the discussion in [1]. It avoids
load ordering dependencies between ice and irdma and allow
'irdma' to be unified without needing to link with the LAN driver.

Summary of changes:
*Remove code that exposes the ice virtual bus and instead register peer devices to platform bus
*Remove peer driver registration/un-registration APIs as the RDMA driver
directly registers with the platform bus.
*Drop probe()/remove() callbacks from driver ops.
*Expose a new callback probe_finished() to notify ice when the RDMA bus probe completes
*Update the peer device state machine.

[1] https://patchwork.kernel.org/patch/10815567/

For the purposes of this RFC, this patch has been kept
separate from the other IDC code so that the changes can
be more easily reviewed; these changes will be incorporated
into the IDC patch series before submission.

Signed-off-by: Dave Ertman <david.m.ertman@intel.com>
Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>
Signed-off-by: Shiraz Saleem <shiraz.saleem@intel.com>