ice: Add support for ice peer devices and drivers
The E800 series of Ethernet devices supports an RDMA hardware block
which isn't interfaced directly to PCI bus. The RDMA driver 'irdma'
thus depends on the ice driver to provide access to the RDMA hardware
block.

The ice driver first creates a pseudo bus and then creates and attaches
a new device to the pseudo bus using device_register(). This new device
is referred to as a "peer device" and the associated driver is a "peer
driver" to ice. Once the peer driver loads, it can call ice driver
functions exposed to it via struct ice_ops. Similarly, ice can call peer
driver exposed to it via struct ice_peer_ops.

The interaction between the peer device and driver is referred to as
Inter-Driver Communication (IDC).

For the purposes of this RFC, this patch is being submitted as a
monolothic patch. It will be broken up into multiple patches for
the final submission.

Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>